code/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64

_libgcc_mutex=0.1=main
absl-py=0.12.0=pypi_0
aiohttp=3.7.4=py37h27cfd23_1
argon2-cffi=20.1.0=py37h27cfd23_1
astunparse=1.6.3=pypi_0
async-timeout=3.0.1=py37h06a4308_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
bidict=0.21.2=pypi_0
biopython=1.78=pypi_0
blas=1.0=mkl
bleach=3.3.0=pyhd3eb1b0_0
boto3=1.17.46=pyhd3eb1b0_0
botocore=1.20.78=pyhd3eb1b0_1
brotlipy=0.7.0=py37h27cfd23_1003
ca-certificates=2021.7.5=h06a4308_1
cachetools=4.2.1=pypi_0
certifi=2021.5.30=py37h06a4308_0
cffi=1.14.5=py37h261ae71_0
chardet=4.0.0=pypi_0
click=8.0.1=pyhd3eb1b0_0
cryptography=3.4.7=py37hd23ed53_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=py37_0
cython=0.29.22=pypi_0
dataclasses=0.6=pypi_0
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=pypi_0
deepsnap=0.2.1=pypi_0
defusedxml=0.7.1=pyhd3eb1b0_0
dgl=0.4.3.post2=pypi_0
easydict=1.9=pypi_0
entrypoints=0.3=py37_0
expat=2.3.0=h2531618_2
flatbuffers=1.12=pypi_0
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
fsspec=2021.10.0=pypi_0
future=0.18.2=pypi_0
gast=0.3.3=pypi_0
gensim=4.0.1=py37h2531618_0
glib=2.68.1=h36276a3_0
google-api-core=1.25.1=pyhd3eb1b0_0
google-auth=1.28.0=pypi_0
google-auth-oauthlib=0.4.3=pypi_0
google-cloud-core=1.6.0=pyhd3eb1b0_0
google-cloud-storage=1.38.0=pyhd3eb1b0_0
google-crc32c=1.1.2=py37h27cfd23_0
google-pasta=0.2.0=pypi_0
google-resumable-media=1.3.0=pyhd3eb1b0_1
googleapis-common-protos=1.53.0=py37h06a4308_0
grakel=0.1.8=pypi_0
grpcio=1.32.0=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=2.10.0=pypi_0
icu=58.2=he6710b0_3
idna=2.10=pyhd3eb1b0_0
imageio=2.22.2=pypi_0
importlib-metadata=3.7.3=pypi_0
importlib_metadata=3.10.0=hd3eb1b0_0
intel-openmp=2020.2=254
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.22.0=py37hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.17.0=py37_0
jinja2=3.0.0=pyhd3eb1b0_0
jmespath=0.10.0=py_0
joblib=0.17.0=pypi_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=py_2
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py37h06a4308_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.1=py37h2531618_0
lap=0.4.0=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.33.1=h53a641e_7
libcrc32c=1.1.1=he6710b0_2
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.14.0=h8c45485_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
line-counter=0.7.4=pypi_0
lz4-c=1.9.2=heb0550a_3
markdown=3.3.4=pypi_0
markupsafe=2.0.0=py37h27cfd23_0
matplotlib=3.3.4=py37h06a4308_0
matplotlib-base=3.3.4=py37h62a2d02_0
mistune=0.8.4=py37h14c3975_1001
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.3.0=py37h54f3939_0
mkl_random=1.1.1=py37h0573a6f_0
multidict=5.1.0=py37h27cfd23_2
nb_conda=2.2.1=py37_0
nb_conda_kernels=2.3.1=py37h06a4308_0
nbconvert=5.6.1=py37_1
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.2=he6710b0_1
networkx=2.6.2=pypi_0
ninja=1.10.1=py37hfd86e86_0
nltk=3.6.2=pyhd3eb1b0_0
nose=1.3.7=pypi_0
notebook=6.3.0=py37h06a4308_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py_0
openssl=1.1.1l=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=20.9=pyhd3eb1b0_0
pandas=1.3.5=pypi_0
pandoc=2.12=h06a4308_0
pandocfilters=1.4.3=py37h06a4308_1
parso=0.8.2=pyhd3eb1b0_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=pypi_0
pip=20.2.4=py37_0
prettytable=2.2.1=pypi_0
prometheus_client=0.10.1=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyh06a4308_0
protobuf=3.13.0=pypi_0
psutil=5.8.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=py_2
pydeprecate=0.3.1=pypi_0
pygments=2.8.1=pyhd3eb1b0_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py37h05f1152_2
pyrsistent=0.17.3=py37h7b6447c_0
pysocks=1.7.1=py37_1
python=3.7.9=h7579374_0
python-dateutil=2.8.1=pyhd3eb1b0_0
python-levenshtein=0.12.2=pypi_0
pytorch=1.7.0=py3.7_cuda10.1.243_cudnn7.6.3_0
pytorch-lightning=1.4.9=pypi_0
pytz=2021.1=pyhd3eb1b0_0
pyyaml=5.3.1=pypi_0
pyzmq=20.0.0=py37h2531618_1
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
regex=2021.8.3=py37h7f8727e_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pyhd3eb1b0_1
s3transfer=0.3.6=pyhd3eb1b0_0
scikit-learn=0.23.2=pypi_0
scipy=1.5.3=pypi_0
seaborn=0.12.1=pypi_0
send2trash=1.5.0=pyhd3eb1b0_1
sentencepiece=0.1.96=pypi_0
setuptools=50.3.0=py37hb0f4dca_1
sip=4.19.8=py37hf484d3e_0
six=1.15.0=py_0
sklearn=0.0=pypi_0
smart_open=5.0.0=pyhd3eb1b0_0
sqlite=3.33.0=h62c20be_0
tensorboard=2.4.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorboardx=2.1=pypi_0
tensorflow=2.4.1=pypi_0
tensorflow-estimator=2.4.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.9.4=py37h06a4308_0
test-tube=0.7.5=pypi_0
testpath=0.4.4=pyhd3eb1b0_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hbc83047_0
torch-cluster=1.5.8=pypi_0
torch-geometric=2.1.0.post1=pypi_0
torch-scatter=2.0.5=pypi_0
torch-sparse=0.6.8=pypi_0
torch-spline-conv=1.2.0=pypi_0
torchaudio=0.7.0=py37
torchmetrics=0.5.1=pypi_0
torchvision=0.8.1=py37_cu101
tornado=6.1=py37h27cfd23_0
tqdm=4.51.0=pypi_0
traitlets=5.0.5=pyhd3eb1b0_0
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.4=pyhd3eb1b0_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py37_1
werkzeug=1.0.1=pypi_0
wheel=0.35.1=py_0
word2vec=0.11.1=pypi_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h7b6447c_0
yarl=1.5.1=py37h7b6447c_0
zeromq=4.3.4=h2531618_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
